\name{corn}
\alias{corn}
\title{Example dataset: corn}
\description{
	Example 1-way mixed-effects ANCOVA dataset for acovamcmc. Classic agricultural experiment: corn yield by variety and level of nitrogen fertilizer.
}
\usage{corn}
\format{
	\describe{
		\item{\code{variety}}{corn vareity}
	}
	\describe{
		\item{\code{nitrogen}}{level of nitrogen fertilizer}
	}
	\describe{
		\item{\code{yield}}{corn yield}
	}
}
\examples{
data(corn)
}

