\name{csection}
\alias{csection}
\title{Example dataset: C-section rates}
\description{
	Example 1-way mixed-effects ANOVA dataset for acovamcmc. C-section rate for 10 randomly selected obstetricians at 12 randomly selected urban hospitals with at least 20 obstetricans associated with its maternity ward.
}
\usage{csection}
\format{
	\describe{
		\item{\code{hospital}}{one of 12 hospitals where C-section rate was recorded}
	}
	\describe{
		\item{\code{rate}}{rate of delivery of babies by C-section}
	}
}
\examples{
data(csection)
}

